create separate pages folders
